requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
kaleido>=0.2.1
bokeh>=3.3.0
